requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
